refactor: IntervalInput 이벤트 및 데이터 훅으로 분리
