+++ frontend/src/components/IntervalInput/IntervalInput.tsx
@@ -1,6 +1,4 @@
-import { useState } from 'react';
-import { minSecToSeconds, secondsToMinSec } from '@/utils/convertTime';
-import { isTimeInSongRange, isValidMinSec } from '@/utils/validateTime';
+import useIntervalInput from './hooks/useIntervalInput';
 import {
   ErrorMessage,
   InputEnd,
@@ -9,52 +7,19 @@ import {
   IntervalContainer,
   Separator,
 } from './IntervalInput.style';
-import { isInputName } from './IntervalInput.type';
-import type { IntervalInputProps, IntervalInputType, TimeMinSec } from './IntervalInput.type';
-import type { ChangeEventHandler, MouseEventHandler } from 'react';
+import type { IntervalInputProps } from './IntervalInput.type';
 
 const IntervalInput = ({ songEnd }: IntervalInputProps) => {
-  const [intervalStart, setIntervalStart] = useState<TimeMinSec>({ minute: 0, second: 0 });
-  const [errorMessage, setErrorMessage] = useState('');
-  const [activeInput, setActiveInput] = useState<IntervalInputType>(null);
-
-  const [endMinute, endSecond] = secondsToMinSec(
-    minSecToSeconds(intervalStart.minute, intervalStart.second) + 10
-  );
-
-  const onChangeIntervalStart: ChangeEventHandler<HTMLInputElement> = ({
-    currentTarget: { name, value },
-  }) => {
-    if (!isValidMinSec(value)) {
-      setErrorMessage('초/분은 0 ~ 59 숫자만 입력 가능해요');
-      return;
-    }
-
-    setErrorMessage('');
-    setIntervalStart({
-      ...intervalStart,
-      [name]: Number(value),
-    });
-  };
-
-  const onClickIntervalStart: MouseEventHandler<HTMLInputElement> = ({
-    currentTarget: { name },
-  }) => {
-    if (isInputName(name)) {
-      setActiveInput(name);
-    }
-  };
-
-  const onBlurIntervalStart = () => {
-    const timeSelected = minSecToSeconds(intervalStart.minute, intervalStart.second);
-
-    if (!isTimeInSongRange({ songEnd, timeSelected })) {
-      const [songMin, songSec] = secondsToMinSec(songEnd - 10);
-      setErrorMessage(`구간 시작을 ${songMin}분 ${songSec}초보다 전으로 설정해주세요`);
-    }
-
-    setActiveInput(null);
-  };
+  const {
+    intervalStart,
+    errorMessage,
+    activeInput,
+    endMinute,
+    endSecond,
+    onChangeIntervalStart,
+    onClickIntervalStart,
+    onBlurIntervalStart,
+  } = useIntervalInput(songEnd);
 
   return (
     <IntervalContainer>
